requirements.txt
Flask
python-docx
pymupdf
sentence-transformers
scikit-learn
torch
gunicorn
